# docs/swagger.yaml
basePath: /
definitions:
...
  models.EmailConfig:
    properties:
      from_email:
        type: string
      from_name:
        type: string
      secure:
        description: TLS/SSL
        type: boolean
      smtp_host:
        type: string
      smtp_password:
        type: string
      smtp_port:
        type: integer
      smtp_user:
        type: string
    type: object
...
  models.Report:
    properties:
      columns:
        description: List of field names to display
        items:
          type: string
        type: array
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      filters:
        additionalProperties: {}
        description: Stored query filters
        type: object
      id:
        type: string
      module_id:
        description: The module this report is based on
        type: string
      name:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
  description: This is a sample server using Fiber, Uber Fx, and GORM.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Microservice Demo API
  version: "1.0"
paths:
...
  /api/modules/{module}/records/{id}:
    get:
      description: Get a record by ID
      parameters:
      - description: Module Name
        in: path
        name: module
        required: true
        type: string
      - description: Record ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get a single record
      tags:
      - records
...
  /api/modules/{name}/records:
    get:
      consumes:
      - application/json
      description: Get records with pagination and filters
      parameters:
      - description: Module Name
        in: path
        name: name
        required: true
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Records per page (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Invalid input
          schema:
            type: string
      summary: List records from a module
      tags:
      - records
    post:
      consumes:
      - application/json
      description: Insert data into a dynamic module with validation
      parameters:
      - description: Module Name
        in: path
        name: name
        required: true
        type: string
      - description: Record Data
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Module not found
          schema:
            type: string
      summary: Create a record in a module
      tags:
      - records
  /api/modules/{name}/records/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a record from a module
      parameters:
      - description: Module Name
        in: path
        name: name
        required: true
        type: string
      - description: Record ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
      summary: Delete a record
      tags:
      - records
    put:
      consumes:
      - application/json
      description: Update a record in a module
      parameters:
      - description: Module Name
        in: path
        name: name
        required: true
        type: string
      - description: Record ID
        in: path
        name: id
        required: true
        type: string
      - description: Record Data
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
      summary: Update a record
      tags:
      - records
...
  /api/reports:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Report'
            type: array
      summary: List all reports
      tags:
      - Reports
    post:
      consumes:
      - application/json
      parameters:
      - description: Report Definition
        in: body
        name: report
        required: true
        schema:
          $ref: '#/definitions/models.Report'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Report'
      summary: Create a new report
      tags:
      - Reports
  /api/reports/{id}:
    delete:
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete a report
      tags:
      - Reports
    get:
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Report'
      summary: Get a report by ID
      tags:
      - Reports
    put:
      consumes:
      - application/json
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      - description: Report Update
        in: body
        name: report
        required: true
        schema:
          $ref: '#/definitions/models.Report'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Report'
      summary: Update a report
      tags:
      - Reports
  /api/reports/{id}/export:
    get:
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Format (default: csv)'
        in: query
        name: format
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: Export a report to CSV
      tags:
      - Reports
  /api/reports/{id}/run:
    get:
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      summary: Run a report and return JSON data
      tags:
      - Reports
...
  /settings/email:
    get:
      description: Get current SMTP settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EmailConfig'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Email Configuration
      tags:
      - settings
    put:
      consumes:
      - application/json
      description: Update SMTP settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update Email Configuration
      tags:
      - settings
...
swagger: "2.0"